# docs/api_surface/OpenAPI.yaml
openapi: 3.0.3
info:
  title: Campus Resource Hub API
  version: 1.0.0
  description: |
    RESTful API for the Campus Resource Hub - A university resource booking and management system.
    
    **Implementation Status:**
    - ✅ existing: Backend implemented and functional
    - ⚠️ partial: Partially implemented
    - ❌ missing: Not yet implemented
    
    **Authentication:** Session-based cookies (Flask-Login)
    
    **Base URL:** `/api`
  contact:
    name: Campus Resource Hub Team
    email: support@campusresourcehub.edu
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://production-domain.com/api
    description: Production server (TBD)
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask-Login session cookie for authentication
    
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF protection token required for all state-changing operations (POST, PUT, PATCH, DELETE).
        Obtain token from GET /api/auth/csrf-token endpoint.
        Token lifetime: 1 hour
        Alternative header name: X-CSRFToken
  schemas:
...
    Review:
      type: object
      properties:
        id:
          type: integer
        resource_id:
          type: integer
        reviewer_id:
          type: integer
        booking_id:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        is_flagged:
          type: boolean
        is_hidden:
          type: boolean
        timestamp:
          type: string
          format: date-time
      required: [id, resource_id, reviewer_id, rating]
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
...
  responses:
...
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per hour
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests (will be 0)
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
        Retry-After:
          schema:
            type: integer
          description: Seconds until retry allowed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Too Many Requests"
              message:
                type: string
                example: "Rate limit exceeded. Please try again later."
              retry_after:
                type: string
                example: "60 seconds"
              status:
                type: integer
                example: 429
paths:
...
  /reviews:
    post:
      tags: [Reviews]
      summary: Submit review
      x-status: existing
      description: |
        Submit a review for a resource.
        Rate limit: 5 per hour
      security:
        - SessionCookie: []
        - CSRFToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resource_id, rating]
              properties:
                resource_id:
                  type: integer
                booking_id:
                  type: integer
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        '201':
          description: Review submitted
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
        '429':
          $ref: '#/components/responses/RateLimitError'
  /resources/{id}/reviews:
    get:
      tags: [Reviews]
      summary: Get resource reviews
      x-status: existing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  average_rating:
                    type: number
                  total_reviews:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...